Fix: YesResultContainer 내 API 2개인 것 분해 위해 SearchResultGameContainer, SearchResultPostContainer 두 컴포넌트 내 각 하나의 API 연결로 변경
