+++ src/container/inNotificationPage/NotificationListContainer.js
@@ -8,7 +8,7 @@ import { Img } from "../../style/ImgStyle";
 import noAlarmImg from "../../img/noAlarmImg.svg";
 import NotificationListItem from "../../component/NotificationListItem";
 
-import { useCookies } from "react-cookie";
+import useFetch from "../../hook/useFetch";
 
 const OverFlowDiv = styled(Section)`
   overflow: auto;
@@ -19,189 +19,70 @@ const NotiListLayout = styled(Article)`
 `;
 
 const NotificationListContainer = () => {
-  const { isAdmin } = { isAdmin: true };
-  const [cookies] = useCookies(["token"]);
+  // const { isAdmin } = { isAdmin: true };
 
-  const [page, setPage] = useState(1);
-  const [adminNotiListData, setAdminNotiListData] = useState([]);
-  console.log("adminNotiListData: ", adminNotiListData);
-  const [notiListData, setNotiListData] = useState([]);
-  console.log("notiListData: ", notiListData);
+  // const [page, setPage] = useState(1);
+  // const [adminNotiListData, setAdminNotiListData] = useState([]);
 
   // 관리자 승인요청 온 게임 목록보기 GET
+  // useEffect(() => {
+  //   const getAdminNotiListData = async () => {
+  //     const response = await fetch(`${process.env.REACT_APP_API_KEY}/admin/game/request/all`, {
+  //       method: "GET",
+  //       headers: {
+  //         "Content-Type": "application/json",
+  //         Authorization: `Bearer ${cookies.token}`,
+  //       },
+  //     });
+
+  //     const result = await response.json();
+
+  //     if (response.status === 200) {
+  //       setAdminNotiListData(result.data);
+  //     } else {
+  //       alert(result.message);
+  //     }
+  //   };
+
+  //   getAdminNotiListData();
+  // }, []);
+
+  // // 관리자 승인요청 온 게임 목록 백엔드 state가 업데이트 될 때 마다, page를 1 증가시키기
+  // useEffect(() => {
+  //   setPage(page + 1);
+  // }, [adminNotiListData]);
+
+  // 일반사용자 알림 목록보기 GET
+
+  const [notiListData, setNotiListData] = useState([]);
+  const [page, setPage] = useState(1);
+
+  const { data, error, status, request } = useFetch();
   useEffect(() => {
-    const getAdminNotiListData = async () => {
-      const response = await fetch(`${process.env.REACT_APP_API_KEY}/admin/game/request/all`, {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${cookies.token}`,
-        },
-      });
-
-      const result = await response.json();
-
-      if (response.status === 200) {
-        setAdminNotiListData(result.data);
-      } else {
-        alert(result.message);
-      }
-    };
-
-    getAdminNotiListData();
-  }, []);
-
-  // 관리자 승인요청 온 게임 목록 백엔드 state가 업데이트 될 때 마다, page를 1 증가시키기
-  useEffect(() => {
-    setPage(page + 1);
-  }, [adminNotiListData]);
-
-  // // 일반사용자 알림 목록보기 GET
-  useEffect(() => {
-    const getNotiListData = async () => {
-      const response = await fetch(`${process.env.REACT_APP_API_KEY}/account/notification`, {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${cookies.token}`,
-        },
-      });
-
-      const result = await response.json();
-
-      if (response.status === 200) {
-        setNotiListData(result.data);
-      } else {
-        alert(result.message);
-      }
-    };
-
-    getNotiListData();
+    request(`/account/notification`, "GET", null);
   }, []);
 
   // 일반사용자 알림 목록 백엔드 state가 업데이트 될 때 마다, page를 1 증가시키기
-  useEffect(() => {
-    setPage(page + 1);
-  }, [notiListData]);
-
-  // 관리자 알림함
   // useEffect(() => {
-  //   setAdminNotiListData([
-  //     {
-  //       id: "Notification_1",
-  //       date: "2024.02.29. 11:01",
-  //       content: "새로운 게임 “(게임 명)” 생성이 요청되었습니다.",
-  //     },
-  //     {
-  //       id: "Notification_2",
-  //       date: "2024.01.29. 11:01",
-  //       content: "새로운 게임 “(게임 명)” 생성이 요청되었습니다.",
-  //     },
-  //     {
-  //       id: "Notification_3",
-  //       date: "2023.12.29. 11:01",
-  //       content: "새로운 게임 “(게임 명)” 생성이 요청되었습니다.",
-  //     },
-  //   ]);
-  // }, []);
-
-  // useEffect(() => {
-  //   const scrollDownEvent = () => {
-  //     const { scrollTop, clientHeight, scrollHeight } = document.documentElement;
-  //     if (scrollTop + clientHeight >= scrollHeight) {
-  //       addAdminNotiDummyData();
-  //     }
-  //   };
-  //   window.addEventListener("scroll", scrollDownEvent);
-  //   return () => {
-  //     window.removeEventListener("scroll", scrollDownEvent);
-  //   };
-  // }, [adminNotiListData]);
-
-  // const addAdminNotiDummyData = () => {
-  //   const newData = [
-  //     {
-  //       id: "Notification_4",
-  //       date: "2023.11.29. 11:01",
-  //       content: "새로운 게임 “(게임 명)” 생성이 요청되었습니다.",
-  //     },
-  //     {
-  //       id: "Notification_5",
-  //       date: "2023.10.29. 11:01",
-  //       content: "새로운 게임 “(게임 명)” 생성이 요청되었습니다.",
-  //     },
-  //     {
-  //       id: "Notification_6",
-  //       date: "2023.09.29. 11:01",
-  //       content: "사용자의 게시글에 새로운 댓글이 달렸습니다. 어서 확인해보세요~!",
-  //     },
-  //     {
-  //       id: "Notification_7",
-  //       date: "2023.08.29. 11:01",
-  //       content: "새로운 게임 “(게임 명)” 생성이 요청되었습니다.",
-  //     },
-  //   ];
-
-  //   setAdminNotiListData((prevData) => [...prevData, ...newData]);
-  // };
-
-  // 일반 사용자 알림함
-  // useEffect(() => {
-  //   setNotiListData([
-  //     {
-  //       id: "Notification_1",
-  //       date: "2024.02.29. 11:01",
-  //       content: "사용자의 “(게임 명)” 게시글에 새로운 댓글이 달렸습니다. 어서 확인해보세요~!",
-  //     },
-  //     {
-  //       id: "Notification_2",
-  //       date: "2024.01.29. 11:01",
-  //       content:
-  //         "사용자가 작성했던 “(게임 명)” 위키가 다른 사용자에 의해 수정되었습니다. 어떻게 바꼈나 궁금하지 않으세요?",
-  //     },
-  //     {
-  //       id: "Notification_3",
-  //       date: "2023.12.29. 11:01",
-  //       content: "요청하신 “(게임 명)” 게임 생성이 거절되었습니다..ㅠㅠ ",
-  //     },
-  //   ]);
-  // }, []);
-
-  // useEffect(() => {
-  //   const scrollDownEvent = () => {
-  //     const { scrollTop, clientHeight, scrollHeight } = document.documentElement;
-  //     if (scrollTop + clientHeight >= scrollHeight) {
-  //       addNotiDummyData();
-  //     }
-  //   };
-  //   window.addEventListener("scroll", scrollDownEvent);
-  //   return () => {
-  //     window.removeEventListener("scroll", scrollDownEvent);
-  //   };
+  //   setPage(page + 1);
   // }, [notiListData]);
 
-  // const addNotiDummyData = () => {
-  //   const newData = [
-  //     {
-  //       id: "Notification_4",
-  //       date: "2023.11.29. 11:01",
-  //       content: "사용자의 “(게임 명)” 게시글에 새로운 댓글이 달렸습니다. 어서 확인해보세요~!",
-  //     },
-  //     {
-  //       id: "Notification_5",
-  //       date: "2023.10.29. 11:01",
-  //       content:
-  //         "사용자가 작성했던 “(게임 명)” 위키가 다른 사용자에 의해 수정되었습니다. 어떻게 바꼈나 궁금하지 않으세요?",
-  //     },
-  //     {
-  //       id: "Notification_6",
-  //       date: "2023.09.29. 11:01",
-  //       content: "요청하신 “(게임 명)” 게임 생성이 거절되었습니다..ㅠㅠ ",
-  //     },
-  //   ];
+  useEffect(() => {
+    if (status === 200) {
+      setNotiListData(data.data);
+    } else if (status === 400) {
+      alert("유효하지 않은 요청입니다.");
+    } else if (status === 401) {
+      alert("권한이 없는 사용자입니다.");
+    } else if (status === 500) {
+      console.log("서버 내부 에러입니다.");
+    }
+  }, [data]);
+  console.log("notiListData: ", notiListData);
 
-  //   setNotiListData((prevData) => [...prevData, ...newData]);
-  // };
+  // if (!data) return "...Loading";
+
+  // if (error) return "Error";
 
   return (
     <OverFlowDiv $height="100%" $flex="v_start_center" $margin="100px 0 0 0" $width="100vw">
