+++ src/container/inSearchResultsPage/SuggestGameContainer.js
@@ -49,7 +49,6 @@ const SuggestGameContainer = (props) => {
         },
         body: JSON.stringify({ title: gameTitle }),
       });
-      // console.log(response.headers);
       const responseContentType = response.headers.get("Content-Type")?.split(";")[0];
       if (responseContentType === "application/json") {
         const result = await response.json();
